+++ src/Compute/Compute/ChangeLog.md
@@ -21,7 +21,7 @@
 -->
 ## Upcoming Release
 * Changed `New-AzVM` cmdlet when using the SimpleParameterSet to not create a PublicIPAddress when a PublicIPAddress name is not provided. 
-* Add `PlatformFaultDomainCounnt` parameter to cmdlets: `New-AzVM` and `New-AzVMConfig`
+* Add `PlatformFaultDomain` parameter to cmdlets: `New-AzVM` and `New-AzVMConfig`
 
 ## Version 4.21.0
 * Contains updates to the following powershell cmdlets